backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
paho-mqtt>=2.0.0
requests
